+++ app/Http/Controllers/UserController.php
@@ -31,10 +31,18 @@ class UserController extends Controller {
     }
 
     public static function getGuildOptions( Request $request, City $city, Guild $guild ) {
+        $user = Auth::user();
+        if( !$user->can('guild_manage', ['guild_id' => $guild->id]) ) {
+            return response()->json('Vous n\'avez pas les permissions nécessaires', 403);
+        }
         return response()->json($guild->settings, 200);
     }
 
     public static function updateGuildOptions( Request $request, City $city, Guild $guild ) {
+        $user = Auth::user();
+        if( !$user->can('guild_manage', ['guild_id' => $guild->id]) ) {
+            return response()->json('Vous n\'avez pas les permissions nécessaires', 403);
+        }
         $settings = $request->settings;
         $guild->updateSettings($settings);
         return response()->json($guild, 200);
@@ -81,11 +89,19 @@ class UserController extends Controller {
     */
 
    public function getConnectors( Request $request, Guild $guild ) {
+       $user = Auth::user();
+       if( !$user->can('guild_manage', ['guild_id' => $guild->id]) ) {
+           return response()->json('Vous n\'avez pas les permissions nécessaires', 403);
+       }
        $connecteurs = Connector::where('guild_id', $guild->id)->get();
        return response()->json($connecteurs, 200);
    }
 
    public function createConnector( Request $request, Guild $guild ) {
+       $user = Auth::user();
+       if( !$user->can('guild_manage', ['guild_id' => $guild->id]) ) {
+           return response()->json('Vous n\'avez pas les permissions nécessaires', 403);
+       }
        $connector = Connector::create([
            'name' => ( isset( $request->name ) ) ? $request->name : '' ,
            'guild_id' => $guild->id,
@@ -104,6 +120,10 @@ class UserController extends Controller {
    }
 
    public function updateConnector( Request $request, Guild $guild, Connector $connector ) {
+       $user = Auth::user();
+       if( !$user->can('guild_manage', ['guild_id' => $guild->id]) ) {
+           return response()->json('Vous n\'avez pas les permissions nécessaires', 403);
+       }
        $connector->update([
            'name' => ( isset( $request->name ) ) ? $request->name : $connector->name ,
            'channel_discord_id' => ( isset( $request->channel_discord_id ) ) ? $request->channel_discord_id : $connector->channel_discord_id ,
@@ -121,10 +141,18 @@ class UserController extends Controller {
    }
 
    public function getConnector( Request $request, Guild $guild, Connector $connector ) {
+       $user = Auth::user();
+       if( !$user->can('guild_manage', ['guild_id' => $guild->id]) ) {
+           return response()->json('Vous n\'avez pas les permissions nécessaires', 403);
+       }
        return response()->json($connector, 200);
    }
 
    public function deleteConnector(Request $request, City $city, Guild $guild, Connector $connector ) {
+       $user = Auth::user();
+       if( !$user->can('guild_manage', ['guild_id' => $guild->id]) ) {
+           return response()->json('Vous n\'avez pas les permissions nécessaires', 403);
+       }
        Connector::destroy($connector->id);
        return response()->json(null, 204);
    }
@@ -136,11 +164,19 @@ class UserController extends Controller {
    */
 
   public function getQuestConnectors( Request $request, Guild $guild ) {
+      $user = Auth::user();
+      if( !$user->can('guild_manage', ['guild_id' => $guild->id]) ) {
+          return response()->json('Vous n\'avez pas les permissions nécessaires', 403);
+      }
       $connecteurs = QuestConnector::where('guild_id', $guild->id)->get();
       return response()->json($connecteurs, 200);
   }
 
   public function createQuestConnector( Request $request, Guild $guild ) {
+      $user = Auth::user();
+      if( !$user->can('guild_manage', ['guild_id' => $guild->id]) ) {
+          return response()->json('Vous n\'avez pas les permissions nécessaires', 403);
+      }
       $connector = QuestConnector::create([
           'name' => ( isset( $request->name ) ) ? $request->name : '' ,
           'guild_id' => $guild->id,
@@ -158,6 +194,10 @@ class UserController extends Controller {
   }
 
   public function updateQuestConnector( Request $request, Guild $guild, QuestConnector $connector ) {
+      $user = Auth::user();
+      if( !$user->can('guild_manage', ['guild_id' => $guild->id]) ) {
+          return response()->json('Vous n\'avez pas les permissions nécessaires', 403);
+      }
       $connector->update([
           'name' => ( isset( $request->name ) ) ? $request->name : $connector->name ,
           'channel_discord_id' => ( isset( $request->channel_discord_id ) ) ? $request->channel_discord_id : $connector->channel_discord_id ,
@@ -174,10 +214,18 @@ class UserController extends Controller {
   }
 
   public function getQuestConnector( Request $request, Guild $guild, QuestConnector $connector ) {
+      $user = Auth::user();
+      if( !$user->can('guild_manage', ['guild_id' => $guild->id]) ) {
+          return response()->json('Vous n\'avez pas les permissions nécessaires', 403);
+      }
       return response()->json($connector, 200);
   }
 
   public function deleteQuestConnector(Request $request, City $city, Guild $guild, QuestConnector $connector ) {
+      $user = Auth::user();
+      if( !$user->can('guild_manage', ['guild_id' => $guild->id]) ) {
+          return response()->json('Vous n\'avez pas les permissions nécessaires', 403);
+      }
       QuestConnector::destroy($connector->id);
       return response()->json(null, 204);
   }
@@ -284,10 +332,6 @@ class UserController extends Controller {
     */
 
     public function getPOIs(City $city, Request $request){
-        $user = Auth::user();
-        if( !$user->can('poi_edit', ['city_id' => $city->id]) ) {
-            return response()->json('Vous n\'avez pas les permissions nécessaires', 403);
-        }
         $pois = Stop::where('city_id', '=', $city->id)
             ->get();
         return response()->json($pois, 200);
